+++ Options/ExternalApiOptions.cs
@@ -1,8 +1,8 @@
 ﻿namespace Prueba_Tecnica_Net.Options
 {
-    public class ApiRetailerOptions
+    public class ExternalApiOptions
     {
-        public const string Key = "ApiRetailers";
+        public const string Key = "ExternalApi";
 
         public string AllRetailersPath { get; set; }
     }
