+++ appsettings.Development.json
@@ -7,7 +7,7 @@
         "ConnectionStrings": {
             "DbConnection": "data source=PruebaTecnicaNet;"
         },
-        "ApiRetailers": {
+        "ExternalApi": {
             "AllRetailersPath": "https://api.opendata.esett.com/EXP01/Retailers"
         }
     }
