+++ appsettings.json
@@ -9,7 +9,7 @@
     "ConnectionStrings": {
         "DbConnection": "data source=PruebaTecnicaNet;"
     },
-    "ApiRetailers": {
+    "ExternalApi": {
         "AllRetailersPath": "https://api.opendata.esett.com/EXP01/Retailers"
     }
 }
